Собрал всё по прокси в отдельный пакет
