+++ app/src/main/res/layout/activity_manual_proxy.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="ru.furry.furview2.UI.ManualProxy">
+    tools:context="ru.furry.furview2.proxy.ManualProxy">
 
     <EditText
         android:layout_width="wrap_content"
